Move Book to mdbook-core
This moves the Book definition to mdbook-core, along with related types
it needs.
