+++ crates/mdbook-summary/Cargo.toml
@@ -10,6 +10,7 @@ rust-version.workspace = true
 [dependencies]
 anyhow.workspace = true
 log.workspace = true
+mdbook-core.workspace = true
 memchr.workspace = true
 pulldown-cmark.workspace = true
 serde.workspace = true
